requirements.txt
requests
python-dotenv
urllib3==1.26.16